Relocating the function `findException`.
